+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Modify FilesystemCompressor options: `FilesystemCompressor::options(&mut self, options: CompressionOptions)`
 
 ## [v0.11.0] - 2023-03-14
 ### Added